merge dev into main

+++ comet/core/database.py
@@ -610,13 +610,14 @@ async def setup_database():
             {"cache_ttl": settings.METADATA_CACHE_TTL, "current_time": time.time()},
         )
 
-        await database.execute(
-            """
-            DELETE FROM torrents
-            WHERE timestamp + :cache_ttl < :current_time;
-            """,
-            {"cache_ttl": settings.TORRENT_CACHE_TTL, "current_time": time.time()},
-        )
+        if settings.TORRENT_CACHE_TTL >= 0:
+            await database.execute(
+                """
+                DELETE FROM torrents
+                WHERE timestamp + :cache_ttl < :current_time;
+                """,
+                {"cache_ttl": settings.TORRENT_CACHE_TTL, "current_time": time.time()},
+            )
 
         await database.execute(
             """
@@ -629,6 +630,8 @@ async def setup_database():
         await database.execute("DELETE FROM download_links_cache")
 
         await database.execute("DELETE FROM active_connections")
+
+        await database.execute("DELETE FROM metrics_cache")
 
     except Exception as e:
         logger.error(f"Error setting up the database: {e}")
